refactor: move dtos folder to @backend folder
